Move the LoginReply type to its own file
